Move services scripts into app/scripts
